Move some tests from MainTestDockerSuite to MainTestNativeSuite
Perform tests that require Docker in the Docker Suite but do others in the Native Suite
Separates functionality a bit more into logical blocks and potentially speeds things up too
